feat: migrate EnvRoute.spec
